reorganize the files to make more sense
