+++ lib/providers/gitee.js
