+++ lib/providers/gitlab.js
